migrated project to postgresql

+++ src/app/services/lobby.service.ts
@@ -1,17 +1,17 @@
 import {HttpException, HttpStatus, Injectable} from '@nestjs/common';
 import {Request as ExpressRequest} from 'express';
 import {LobbyDTO} from '../dto/responses/lobby.dto';
-import {IUser} from '../business/interfaces/user.interface';
-import {UserDao} from '../business/dao/user.dao';
-import {ILobby} from '../business/interfaces/lobby.interface';
-import {LobbyDao} from '../business/dao/lobby.dao';
+import {UserDao} from '../database/dao/user.dao';
+import {LobbyDao} from '../database/dao/lobby.dao';
 import {ChildProcess, fork} from 'child_process';
 import * as path from 'path';
 import {GamingServersManager} from '../managers/gaming-servers.manager';
 import {ChildProcessMessage} from '../../gaming-worker/consts/child-process-message.const';
 import {AppUtils} from '../utils/app.utils';
-import {GameState} from '../business/interfaces/enum/game-state.enum';
 import {LobbyInitialDataDTO} from '../dto/requests/lobby-initial-data.dto';
+import LobbyEntity from '../database/entities/lobby.entity';
+import {GameState} from '../database/datatypes/game-state.enum';
+import UserEntity from '../database/entities/user.entity';
 
 @Injectable()
 export class LobbyService {
@@ -21,12 +21,12 @@ export class LobbyService {
     }
 
     async quickStart(req: ExpressRequest): Promise<LobbyDTO> {
-        const openLobbies: ILobby[] = await this.lobbyDao.getOpenedLobbies();
-        let randomLobby: ILobby;
+        const openLobbies: LobbyEntity[] = await this.lobbyDao.getOpenedLobbies();
+        let randomLobby: LobbyEntity;
         while (!randomLobby && openLobbies.length > 0) {
             randomLobby = openLobbies[ Math.floor(Math.random() * openLobbies.length) ];
             const lobbyPort: number = +randomLobby.serverUrl.substr(randomLobby.serverUrl.lastIndexOf(':') + 1);
-            const lobbyId: string = this.gamingServersManager.getRunningLobbyId(lobbyPort);
+            const lobbyId: number = this.gamingServersManager.getRunningLobbyId(lobbyPort);
             if ( !lobbyId || lobbyId != randomLobby.id ) {
                 console.log(`Found orphan lobby (port ${lobbyPort})! ${lobbyId ? ('Expected ' + randomLobby.id + '; found ' + lobbyId) : 'There is no such lobby in gaming server manager'}`);
                 randomLobby.state = GameState.Errored;
@@ -42,14 +42,17 @@ export class LobbyService {
     }
 
     async getLobbies(): Promise<LobbyDTO[]> {
-        const lobbies: ILobby[] = await this.lobbyDao.getOpenedLobbies();
-        return lobbies.map((lobby: ILobby): LobbyDTO => LobbyDTO.fromLobby(lobby));
+        const lobbies: LobbyEntity[] = await this.lobbyDao.getOpenedLobbies();
+        return lobbies.map((lobby: LobbyEntity): LobbyDTO => LobbyDTO.fromLobby(lobby));
     }
 
     async createLobby(req: ExpressRequest, initialData: LobbyInitialDataDTO = new LobbyInitialDataDTO()): Promise<LobbyDTO> {
-        const userModel: IUser = await this.userDao.getUserById(req.session.user._id);
-        const lobbyModel: ILobby = await this.lobbyDao.create(userModel);
-        let port: number = this.gamingServersManager.acquirePort(lobbyModel._id);
+        const userModel: UserEntity = await this.userDao.getUserById(req.session.user.id);
+        if (!userModel) {
+            throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
+        }
+        const lobbyModel: LobbyEntity = await this.lobbyDao.create(userModel);
+        let port: number = this.gamingServersManager.acquirePort(lobbyModel.id);
         if ( port === -1 ) {
             await this.lobbyDao.delete(lobbyModel);
             throw new HttpException('Server overloaded. Try again later', HttpStatus.NOT_ACCEPTABLE);
@@ -79,27 +82,27 @@ export class LobbyService {
                 }
             });
             gamingProcess.on('close', () => {
-                console.log(`Gaming server for lobby ${lobbyModel._id} closed`);
-                this.gamingServersManager.releasePort(lobbyModel._id);
+                console.log(`Gaming server for lobby ${lobbyModel.id} closed`);
+                this.gamingServersManager.releasePort(lobbyModel.id);
             });
             return LobbyDTO.fromLobby(lobbyModel);
         } catch (err) {
             if ( gamingProcess ) {
                 gamingProcess.kill();
             }
-            this.gamingServersManager.releasePort(lobbyModel._id);
+            this.gamingServersManager.releasePort(lobbyModel.id);
             throw new HttpException('Gaming server start failed. Try again later', HttpStatus.GONE);
         }
     }
 
-    async startGamingServer(lobbyModel: ILobby, port: number, owner: IUser): Promise<ChildProcess> {
+    async startGamingServer(lobbyModel: LobbyEntity, port: number, owner: UserEntity): Promise<ChildProcess> {
         // automatically rejects in 10 seconds if nothing happen
         return new Promise<ChildProcess>(async (resolve, reject) => {
             const forked: ChildProcess = fork(path.join(__dirname, '../../../index-worker.js'),
                 [
-                    lobbyModel._id,
+                    lobbyModel.id,
                     port,
-                    owner._id,
+                    owner.id,
                     lobbyModel.playersCount,
                 ],
                 AppUtils.isInDebugMode() ?
